Refactor: Moved NewsletterSubscribeInput into separate component file
